+++ src/components/pages/SettingsPage.tsx
@@ -43,9 +43,6 @@ import { FixedSpacer } from "ui/spacing/Spacing";
 import { useAppDispatch, useAppSelector } from "store/hooks";
 import { ColorModeSelect } from "components/forms/ColorModeSelect";
 import { CompilerPresetSelect } from "components/forms/CompilerPresetSelect";
-import { CompilerOptimisationSelect } from "components/forms/CompilerOptimisationSelect";
-import { CompilerOptimisation } from "shared/lib/resources/types";
-import Alert from "ui/alerts/Alert";
 
 const SettingsPage: FC = () => {
   const dispatch = useAppDispatch();
@@ -89,7 +86,6 @@ const SettingsPage: FC = () => {
     musicDriver,
     openBuildLogOnWarnings,
     generateDebugFilesEnabled,
-    compilerOptimisation,
     compilerPreset,
   } = settings;
 
@@ -156,12 +152,6 @@ const SettingsPage: FC = () => {
   const onChangeGenerateDebugFilesEnabled = useCallback(
     (e: React.ChangeEvent<HTMLInputElement>) =>
       onChangeSettingProp("generateDebugFilesEnabled", castEventToBool(e)),
-    [onChangeSettingProp]
-  );
-
-  const onChangeCompilerOptimisation = useCallback(
-    (value: CompilerOptimisation) =>
-      onChangeSettingProp("compilerOptimisation", value),
     [onChangeSettingProp]
   );
 
@@ -670,7 +660,6 @@ const SettingsPage: FC = () => {
             l10n("FIELD_OPEN_BUILD_LOG_ON_WARNINGS"),
             l10n("FIELD_GENERATE_DEBUG_FILES"),
             l10n("FIELD_COMPILER_PRESET"),
-            l10n("FIELD_COMPILER_OPTIMISATION"),
           ]}
         >
           <CardAnchor id="settingsBuild" />
@@ -711,22 +700,6 @@ const SettingsPage: FC = () => {
 
           <SearchableSettingRow
             searchTerm={searchTerm}
-            searchMatches={[l10n("FIELD_COMPILER_OPTIMISATION")]}
-          >
-            <SettingRowLabel>
-              {l10n("FIELD_COMPILER_OPTIMISATION")}
-            </SettingRowLabel>
-            <SettingRowInput>
-              <CompilerOptimisationSelect
-                name={"compilerOptimisation"}
-                value={compilerOptimisation}
-                onChange={onChangeCompilerOptimisation}
-              />
-            </SettingRowInput>
-          </SearchableSettingRow>
-
-          <SearchableSettingRow
-            searchTerm={searchTerm}
             searchMatches={[l10n("FIELD_COMPILER_PRESET")]}
           >
             <SettingRowLabel>{l10n("FIELD_COMPILER_PRESET")}</SettingRowLabel>
@@ -736,11 +709,6 @@ const SettingsPage: FC = () => {
                 value={compilerPreset}
                 onChange={onChangeCompilerPreset}
               />
-              {compilerOptimisation !== "none" && compilerPreset !== 3000 && (
-                <Alert variant="warning" style={{ marginTop: 3 }}>
-                  <p>{l10n("FIELD_COMPILER_OPTIONS_WARNING")}</p>
-                </Alert>
-              )}
             </SettingRowInput>
           </SearchableSettingRow>
 
@@ -750,7 +718,6 @@ const SettingsPage: FC = () => {
                 onClick={() => {
                   onChangeSettingProp("openBuildLogOnWarnings", true);
                   onChangeSettingProp("generateDebugFilesEnabled", false);
-                  onChangeSettingProp("compilerOptimisation", "none");
                   onChangeSettingProp("compilerPreset", 3000);
                 }}
               >