+++ src/lib/compiler/buildMakeScript.ts
@@ -3,7 +3,6 @@ import { promisify } from "util";
 import { pathExists, readFile, writeFile } from "fs-extra";
 import Path from "path";
 import l10n from "shared/lib/lang/l10n";
-import { CompilerOptimisation } from "shared/lib/resources/types";
 
 const globAsync = promisify(glob);
 
@@ -16,7 +15,6 @@ type BuildOptions = {
   batteryless: boolean;
   targetPlatform: "gb" | "pocket";
   cartType: "mbc3" | "mbc5";
-  compilerOptimisation: CompilerOptimisation;
   compilerPreset: number;
 };
 
@@ -123,7 +121,6 @@ export const getBuildCommands = async (
     targetPlatform,
     cartType,
     compilerPreset,
-    compilerOptimisation,
   }: BuildOptions
 ) => {
   const srcRoot = `${buildRoot}/src/**/*.@(c|s)`;
@@ -158,12 +155,6 @@ export const getBuildCommands = async (
       ];
 
       buildArgs.push(`-Wf"--max-allocs-per-node ${compilerPreset ?? 3000}"`);
-
-      if (compilerOptimisation === "speed") {
-        buildArgs.push("-Wf--opt-code-speed");
-      } else if (compilerOptimisation === "size") {
-        buildArgs.push("-Wf--opt-code-size");
-      }
 
       if (colorEnabled) {
         buildArgs.push("-DCGB");